% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cga}
\alias{cga}
\title{Compact Genetic Algorithm}
\usage{
cga(chsize, popsize, evalFunc)
}
\arguments{
\item{chsize}{Number of bits.}

\item{popsize}{Size of population. The value is used for mutating 
the probability vector by 1/popsize.}

\item{evalFunc}{Objective function.}
}
\value{
Binary vector of size chsize.
}
\description{
Performs a Compact Genetic Algorithm (CGA) search
	for a given chromosome size, population size (mutation rate), 
    and an objective function.
}
