\name{eive-package}
\alias{eive-package}
\alias{eive}
\docType{package}
\title{
EIVE - Errors-in-Variable estimation
}
\description{
This package includes functions for compact genetic algorithms and errors-in-variable estimation. The function 'eive' performs a genetic search to reduce the errors-in-variable bias in ordinary least squares estimator. 
}
\details{
\tabular{ll}{
Package: \tab eive\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-02\cr
License: \tab GPL\cr
}
}
\author{
Mehmet Hakan Satman <mhsatman@istanbul.edu.tr>
Erkin Diyarbakirlioglu <ediyarbakirlioglu@gmail.com>


Maintainer: Mehmet Hakan Satman <mhsatman@istanbul.edu.tr>
}
