\name{f}
\alias{f}
\title{
Cost function of cga in errors-in-variables estimation (eive).
}
\description{
This function is not directly called by the user.
}
\usage{
f(x)
}
\arguments{
  \item{x}{
Candidate dummies required by the cost function of CGA.
}
}
\details{
This function is not directly called by the user.
}
\value{
Returns a cost value for a candidate solution in CGA.
}
\author{
Mehmet Hakan Satman <mhsatman@istanbul.edu.tr>
Erkin Diyarbakirlioglu <ediyarbakirlioglu@gmail.com>
}


