% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eive.R
\name{generate.eive.data}
\alias{generate.eive.data}
\title{Generates simulated errors-in-variables regression data}
\usage{
generate.eive.data(n, e.sd, delta.sd, seed = 12345, useotherx = FALSE)
}
\arguments{
\item{n}{Number of observations}

\item{e.sd}{Standard deviation of error term of regression}

\item{delta.sd}{Standard deviation of error in exploratory variable}

\item{seed}{Seed for random number generator. 12345 by default}

\item{useotherx}{Logical. If TRUE, an additional independent variable is added.}
}
\value{
A matrix of variables.
}
\description{
Generates simulated errors-in-variables regression data
}
\section{Slots}{

\describe{
\item{\code{xdelta}}{Errorenous X variable}

\item{\code{otherx}}{Other X variable}

\item{\code{y}}{Response variable}
}}

