% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cga_generate_chromosome}
\alias{cga_generate_chromosome}
\title{Generate Chromosome}
\usage{
cga_generate_chromosome(prob_vec, vect)
}
\arguments{
\item{prob_vec}{Vector of probabilities}

\item{vect}{Vector of bits.}
}
\value{
Mutates the vect. Returns null.
}
\description{
Generate a binary vector using a probability vector
	This function is not directly called by user. CGAs (Compact genetic algorithms)
	sample chromosomes using this probability vector. A probability vector
	contains[P1, P2, ..., PN] and the function generates and returns a chromosome[B1, B2, ..., BN].
	The probability of BK having the value of 1 is PK. So, it has more chance to have
	[1, 1, 1, 0, 0] rather than [0, 0, 0, 1, 1] when the probability vector is
 	[0.9, 0.9, 0.9, 0.1, 0.1].
}
